+++ Posh-ACME/Public/Get-PAAccount.ps1
@@ -23,7 +23,8 @@ function Get-PAAccount {
     Begin {
         # make sure we have a server configured
         if (!(Get-PAServer)) {
-            throw "No ACME server configured. Run Set-PAServer first."
+            try { throw "No ACME server configured. Run Set-PAServer first." }
+            catch { $PSCmdlet.ThrowTerminatingError($_) }
         }
 
         # make sure the Contact emails have a "mailto:" prefix