+++ Posh-ACME/Public/Get-PAPluginArgs.ps1
@@ -7,11 +7,10 @@ function Get-PAPluginArgs {
     )
 
     Begin {
-        trap { $PSCmdlet.ThrowTerminatingError($PSItem) }
-
         # Make sure we have an account configured
         if (-not ($acct = Get-PAAccount)) {
-            throw "No ACME account configured. Run Set-PAAccount or New-PAAccount first."
+            try { throw "No ACME account configured. Run Set-PAAccount or New-PAAccount first." }
+            catch { $PSCmdlet.ThrowTerminatingError($_) }
         }
     }